+++ MainGame/drawables/Tilemap.cpp
@@ -15,10 +15,10 @@ void Tilemap::mutableUpdateVertexMap(sf::Transform transform) const
     auto invTransform = transform.getInverse();
     auto targetFrame = invTransform.transformRect(drawingFrame);
 
-    size_t width = (size_t)floorf((targetFrame.left + targetFrame.width)/tileSize) -
-                   (size_t)floorf((targetFrame.left)/tileSize) + 1;
-    size_t height = (size_t)floorf((targetFrame.top + targetFrame.height)/tileSize) -
-                    (size_t)floorf((targetFrame.top)/tileSize) + 1;
+    uintmax_t width = (uintmax_t)floorf((targetFrame.left + targetFrame.width)/tileSize) -
+                   (uintmax_t)floorf((targetFrame.left)/tileSize) + 1;
+    uintmax_t height = (uintmax_t)floorf((targetFrame.top + targetFrame.height)/tileSize) -
+                    (uintmax_t)floorf((targetFrame.top)/tileSize) + 1;
 
     bool dirty = false;
 
@@ -29,8 +29,8 @@ void Tilemap::mutableUpdateVertexMap(sf::Transform transform) const
         lastPoint = sf::Vector2i(0, 0);
         dirty = true;
 
-        for (size_t j = 0; j < height; j++)
-            for (size_t i = 0; i < width; i++)
+        for (uintmax_t j = 0; j < height; j++)
+            for (uintmax_t i = 0; i < width; i++)
             {
                 vertices[(j*width+i)*4+0].color = sf::Color::White;
                 vertices[(j*width+i)*4+0].position = sf::Vector2f(i*tileSize, j*tileSize);
@@ -46,14 +46,14 @@ void Tilemap::mutableUpdateVertexMap(sf::Transform transform) const
             }
     }
 
-    sf::Vector2i pt(size_t(targetFrame.left/tileSize), size_t(targetFrame.top/tileSize));
+    sf::Vector2i pt(uintmax_t(targetFrame.left/tileSize), uintmax_t(targetFrame.top/tileSize));
 
     if (dirty || lastPoint != pt)
     {
-        size_t stride = texture->getSize().x / tileSize;
+        uintmax_t stride = texture->getSize().x / tileSize;
 
-        for (size_t j = 0; j < height; j++)
-            for (size_t i = 0; i < width; i++)
+        for (uintmax_t j = 0; j < height; j++)
+            for (uintmax_t i = 0; i < width; i++)
             {
                 auto cur = pt + sf::Vector2i(i, j);
 
@@ -66,9 +66,9 @@ void Tilemap::mutableUpdateVertexMap(sf::Transform transform) const
                 }
                 else
                 {
-                    size_t data = tileData(cur.x, cur.y);
-                    size_t texS = data % stride;
-                    size_t texT = data / stride;
+                    uintmax_t data = tileData(cur.x, cur.y);
+                    uintmax_t texS = data % stride;
+                    uintmax_t texT = data / stride;
 
                     vertices[(j*width+i)*4+0].texCoords = sf::Vector2f(tileSize * texS, tileSize * texT);
                     vertices[(j*width+i)*4+1].texCoords = sf::Vector2f(tileSize * (texS+1), tileSize * texT);
