+++ MainGame/objects/RoomShapeGenerator.cpp
@@ -15,21 +15,21 @@ std::vector<std::shared_ptr<cp::Shape>>
     const auto& layer = data.mainLayer;
 
     enum NodeType { None, Knee, Ankle };
-    struct HorSegment { ssize_t y, x1, x2; };
-    struct VertSegment { ssize_t x, y1, y2; };
+    struct HorSegment { intmax_t y, x1, x2; };
+    struct VertSegment { intmax_t x, y1, y2; };
 
-    auto isSolid = [&](ssize_t x, ssize_t y)
+    auto isSolid = [&](intmax_t x, intmax_t y)
     {
         return tileSet.tileModes[std::min<uint8_t>(layer(x, y),
             tileSet.tileModes.size()-1)] == TileSet::Mode::Solid;
     };
-    auto horizontalFootprint = [&](ssize_t j, ssize_t i) -> ssize_t
+    auto horizontalFootprint = [&](intmax_t j, intmax_t i) -> intmax_t
     {
-        return (ssize_t(isSolid(i, j-1)) << 1) | ssize_t(isSolid(i, j));
+        return (intmax_t(isSolid(i, j-1)) << 1) | intmax_t(isSolid(i, j));
     };
-    auto verticalFootprint = [&](ssize_t j, ssize_t i) -> ssize_t
+    auto verticalFootprint = [&](intmax_t j, intmax_t i) -> intmax_t
     {
-        return (ssize_t(isSolid(j-1, i)) << 1) | ssize_t(isSolid(j, i));
+        return (intmax_t(isSolid(j-1, i)) << 1) | intmax_t(isSolid(j, i));
     };
 
     std::vector<HorSegment> horizontalSegments;
@@ -38,9 +38,9 @@ std::vector<std::shared_ptr<cp::Shape>>
 /*#define GEN_SEGMENTS(perp, par, dir)                                                 \
     do                                                                               \
     {                                                                                \
-        for (ssize_t j = 1; j < layer.perp(); j++)                                   \
+        for (intmax_t j = 1; j < layer.perp(); j++)                                   \
         {                                                                            \
-            ssize_t i = 0;                                                           \
+            intmax_t i = 0;                                                           \
                                                                                      \
             while (i < layer.par())                                                  \
             {                                                                        \
@@ -60,9 +60,9 @@ std::vector<std::shared_ptr<cp::Shape>>
 
     do
     {
-        for (ssize_t j = 1; j < layer.height(); j++)
+        for (intmax_t j = 1; j < layer.height(); j++)
         {
-            ssize_t i = 0;
+            intmax_t i = 0;
             while (i < layer.width())
             {
                 auto previ = i;
@@ -76,9 +76,9 @@ std::vector<std::shared_ptr<cp::Shape>>
     
     do
     {
-        for (ssize_t j = 1; j < layer.width(); j++)
+        for (intmax_t j = 1; j < layer.width(); j++)
         {
-            ssize_t i = 0;
+            intmax_t i = 0;
             while (i < layer.height())
             {
                 auto previ = i;
@@ -105,7 +105,7 @@ std::vector<std::shared_ptr<cp::Shape>>
                 (isDown ? -params.perpnOffset + params.cornerRadius                                   \
                         :  params.perppOffset - params.cornerRadius);                                 \
                                                                                                       \
-            if (seg.i##1 == 0) endPoint1.i = -(ssize_t)DefaultTileSize;                               \
+            if (seg.i##1 == 0) endPoint1.i = -(intmax_t)DefaultTileSize;                               \
             else                                                                                      \
             {                                                                                         \
                 endPoint1.i = DefaultTileSize * seg.i##1;                                             \
@@ -128,11 +128,11 @@ std::vector<std::shared_ptr<cp::Shape>>
         }                                                                                             \
     } while (0)
 
-    size_t bounds = layer.width();
+    uintmax_t bounds = layer.width();
     CONVERT_SHAPES(y, x, upperOffset, lowerOffset, leftOffset, rightOffset, horizontalSegments, isSolid);
 
     bounds = layer.height();
-    auto isReverseSolid = [&](ssize_t i, ssize_t j) { return isSolid(j, i); };
+    auto isReverseSolid = [&](intmax_t i, intmax_t j) { return isSolid(j, i); };
     CONVERT_SHAPES(x, y, leftOffset, rightOffset, upperOffset, lowerOffset, verticalSegments, isReverseSolid);
 
     return shapes;
