+++ MainGame/objects/collectibles/Powerup.hpp
@@ -16,7 +16,7 @@ namespace collectibles
         sf::VertexArray vertices;
         std::shared_ptr<sf::Texture> texture;
         std::shared_ptr<cp::Shape> collisionShape;
-        size_t abilityLevel;
+        uintmax_t abilityLevel;
 
         float rotationAngle;
         void makeRotatedRect();
